COM-3309 delete traineeFollowUpTableMixin
